requirements.txt
aniso8601==3.0.2
asn1crypto==0.24.0
beautifulsoup4==4.6.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3.1
Flask==1.0.2
Flask-HTTPAuth==3.2.4
Flask-MySQL==1.4.0
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
lxml==4.2.4
MarkupSafe==1.0
mysql-connector-python==8.0.12
numpy==1.15.0
pandas==0.23.4
protobuf==3.6.1
pycparser==2.19
PyMySQL==0.9.2
python-dateutil==2.7.3
python-dotenv==0.9.1
pytz==2018.5
requests==2.19.1
selenium==3.14.0
six==1.11.0
SQLAlchemy==1.2.12
urllib3==1.23
Werkzeug==0.14.1
